[BTAT-2248] Mirror core model package structure in Unit Tests
