+++ test/models/core/CessationModelSpec.scala
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package models
+package models.core
 
 import java.time.LocalDate
 
 import assets.CessationTestConstants._
-import models.core.CessationModel
 import org.scalatest.Matchers
 import play.api.libs.json._
 import utils.TestSupport
